Improve TeamDao and TeamDaoTest

+++ persistence/src/main/java/cz/muni/fi/dao/TeamDaoImpl.java
@@ -20,15 +20,25 @@ public class TeamDaoImpl implements TeamDao {
 
     public Team addTeam(Team team) {
         if(team == null){
-            throw new IllegalArgumentException("argument is null");
+            throw new IllegalArgumentException("Team argument is null!");
         }
+        if (team.getName() == null || team.getCarOne() == null || team.getCarTwo() == null) {
+            throw new IllegalArgumentException("One of team attributes is null!");
+        }
+        if (team.getId() != null) {
+            throw new IllegalArgumentException("Team id must be null before storing!");
+        }
+
         em.persist(team);
         return team;
     }
 
     public Team updateTeam(Team team) {
         if(team == null){
-            throw new IllegalArgumentException("argument is null");
+            throw new IllegalArgumentException("Team argument is null!");
+        }
+        if (team.getName() == null || team.getCarOne() == null || team.getCarTwo() == null || team.getId() == null) {
+            throw new IllegalArgumentException("One of team attributes is null!");
         }
         em.merge(team);
         return team;
@@ -36,7 +46,10 @@ public class TeamDaoImpl implements TeamDao {
 
     public boolean deleteTeam(Team team) {
         if(team == null){
-            throw new IllegalArgumentException("argument is null");
+            throw new IllegalArgumentException("Team argument is null!");
+        }
+        if (team.getName() == null || team.getCarOne() == null || team.getCarTwo() == null || team.getId() == null) {
+            throw new IllegalArgumentException("One of team attributes is null!");
         }
         em.remove(em.merge(team));
         return true;
@@ -48,8 +61,9 @@ public class TeamDaoImpl implements TeamDao {
 
     public Team findTeamByName(String name) {
         if(name == null){
-            throw new IllegalArgumentException("argument is null");
+            throw new IllegalArgumentException("Name argument cannot is null!");
         }
+
         List<Team> result = em.createQuery("SELECT t FROM Team t WHERE t.name = :name", Team.class)
                             .setParameter("name", name).getResultList();
         if(result.isEmpty()){
@@ -60,7 +74,7 @@ public class TeamDaoImpl implements TeamDao {
 
     public Team findTeamById(Long id) {
         if(id == null){
-            throw new IllegalArgumentException("argument is null");
+            throw new IllegalArgumentException("Id argument is null!");
         }
         return em.find(Team.class, id);
     }
